Removed the auto add by convention
